some fix_add scenes

+++ srcs/read_init/valid_help.c
@@ -6,7 +6,7 @@
 /*   By: gwaymar- <gwaymar-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/25 03:39:41 by gwaymar-          #+#    #+#             */
-/*   Updated: 2019/10/25 04:10:01 by gwaymar-         ###   ########.fr       */
+/*   Updated: 2019/10/25 05:44:59 by gwaymar-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ double		valid_intens(double n)
 	else if (n < 0)
 		n = 0;
 	else
-		n = n;
+		n = (double)n;
 	return (n);
 }
 
@@ -28,7 +28,7 @@ double		valid_radius(double n)
 	if (n < 0)
 		n = 0;
 	else
-		n = n;
+		n = (double)n;
 	return (n);
 }
 
@@ -37,6 +37,6 @@ double		valid_blesk(double n)
 	if (n < 0)
 		n = 0;
 	else
-		n = n;
+		n = (double)n;
 	return (n);
 }
